refactorizo metodo extraerDinero
