+++ src/main/java/com/example/cajeroautomatico/Cuenta.java
@@ -1,24 +1,22 @@
 package com.example.cajeroautomatico;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Scanner;
 
 public abstract class Cuenta {
-        Integer saldo;
+        Integer saldoActual;
+        Integer saldoMinimo;
         ArrayList<Integer> opciones;
 
     public Cuenta(){
         opciones = new ArrayList<>();
-        saldo = 0;
-        opciones.add(5000);
-        opciones.add(10000);
-        opciones.add(20000);
-        opciones.add(50000);
-        opciones.add(100000);
+        saldoActual = 0;
+        opciones.addAll(Arrays.asList(5000, 10000, 20000, 50000, 10000));
     }
 
     public void mostrarSaldo() {
-        System.out.println("Su saldo actual es: " + saldo);
+        System.out.println("Su saldo actual es: " + saldoActual);
     }
 
     public void depositarValor(TipoDeMoneda tipo){
@@ -26,23 +24,30 @@ public abstract class Cuenta {
         Scanner in = new Scanner(System.in);
         System.out.println("Ingrese una cantidad a depositar: ");
         cantidadADepositar = in.nextInt();
-        saldo += cantidadADepositar;
+        saldoActual += cantidadADepositar;
         System.out.println("Su deposito es es de: " + cantidadADepositar);
-        System.out.println("Su saldo actual es: " + saldo);
+        mostrarSaldo();
     }
 
-    public abstract void extraerDinero();
-
+    public void extraerDinero() {
+        Scanner in = new Scanner(System.in);
+        int opcionElegida;
+        mostrarOpciones(this.opciones);
+        opcionElegida = in.nextInt();
+        if(saldoActual - opciones.get(opcionElegida) > saldoMinimo){
+            saldoActual -= opciones.get(opcionElegida);
+        }else {
+            System.out.println("Saldo insuficiente");
+        }
+        mostrarSaldo();
+    }
     public abstract String getInfo();
 
+    //Esto capaz deberia estar en la clase cajero
     public void mostrarOpciones(ArrayList<Integer> opciones) {
-
         System.out.println("Seleccione un monto a extraer:");
         for(int i = 0; i < opciones.size(); i++) {
             System.out.println("Opcion: " + (i+1) + ": " + opciones.get(i));
-
         }
-
-
     }
 }
